fix(np/pt): explicit dtype and device. (#4241)
fix: https://github.com/deepmodeling/deepmd-kit/issues/4016

+++ deepmd/utils/data.py
@@ -266,7 +266,7 @@ class DeepmdData:
         iterator_1 = self.iterator + batch_size
         if iterator_1 >= set_size:
             iterator_1 = set_size
-        idx = np.arange(self.iterator, iterator_1)  # pylint: disable=no-explicit-dtype
+        idx = np.arange(self.iterator, iterator_1, dtype=np.int64)
         self.iterator += batch_size
         ret = self._get_subdata(self.batch_set, idx)
         return ret
@@ -290,7 +290,7 @@ class DeepmdData:
                 else self.test_set["type"].shape[0]
             )
             # print('ntest', self.test_set['type'].shape[0], ntests, ntests_)
-            idx = np.arange(ntests_)  # pylint: disable=no-explicit-dtype
+            idx = np.arange(ntests_, dtype=np.int64)
         ret = self._get_subdata(self.test_set, idx=idx)
         if self.modifier is not None:
             self.modifier.modify_data(ret, self)
@@ -378,14 +378,14 @@ class DeepmdData:
                 new_types.append(ii)
         new_types = np.array(new_types, dtype=int)
         natoms = new_types.shape[0]
-        idx = np.arange(natoms)  # pylint: disable=no-explicit-dtype
+        idx = np.arange(natoms, dtype=np.int64)
         idx_map = np.lexsort((idx, new_types))
         return idx_map
 
     def _get_natoms_2(self, ntypes):
         sample_type = self.atom_type
         natoms = len(sample_type)
-        natoms_vec = np.zeros(ntypes).astype(int)  # pylint: disable=no-explicit-dtype
+        natoms_vec = np.zeros(ntypes, dtype=np.int64)
         for ii in range(ntypes):
             natoms_vec[ii] = np.count_nonzero(sample_type == ii)
         return natoms, natoms_vec
@@ -435,7 +435,7 @@ class DeepmdData:
     def _shuffle_data(self, data):
         ret = {}
         nframes = data["coord"].shape[0]
-        idx = np.arange(nframes)  # pylint: disable=no-explicit-dtype
+        idx = np.arange(nframes, dtype=np.int64)
         # the training times of each frame
         idx = np.repeat(idx, np.reshape(data["numb_copy"], (nframes,)))
         dp_random.shuffle(idx)
@@ -676,7 +676,7 @@ class DeepmdData:
 
     def _make_idx_map(self, atom_type):
         natoms = atom_type.shape[0]
-        idx = np.arange(natoms)  # pylint: disable=no-explicit-dtype
+        idx = np.arange(natoms, dtype=np.int64)
         if self.sort_atoms:
             idx_map = np.lexsort((idx, atom_type))
         else:
